Improved migration scripImproved migration scriptt
