+++ packs/heroMacros/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000476
+MANIFEST-000502
