+++ packs/heroMacros/LOG
@@ -1,3 +1,1 @@
-2023/09/02-08:34:41.096 6fbc Recovering log #475
-2023/09/02-08:34:41.101 6fbc Delete type=0 #475
-2023/09/02-08:34:41.101 6fbc Delete type=3 #474
+2023/09/03-13:46:01.882 4f5c Recovering log #500
