+++ templates/actor-sidebar/actor-sidebar-sheet.hbs
@@ -78,10 +78,8 @@
                         <button type="button" class="full-health-button"
                             title="Restore all characteristics to max, remove status effects, and remote temporary effects">Full
                             Health</button>
-                        {{#if isGM}}
                         <button type="button" class="actor-description-button"
                             title="Show Actor Description and all percievable powers">Actor Description</button>
-                        {{/if}}
                     </div>
 
             </section>
